Created handler function Fixes #3
